Reorganize analysis folder structure without code changes
- Move processor files to src/analysis/processors/ directory
- Update import paths in processor.py
- Add TODO file for future improvements
- Maintain backward compatibility with existing code
